added header and definition to prevent redefinition linker issues
